完成GIF控件调整

+++ src/LayuiAvaloniaPack/UI/LayUI.Avalonia/Controls/Image/GIF/Extensions/StreamExtensions.cs
@@ -48,7 +48,7 @@ namespace LayUI.Avalonia.Controls.GIF.Extensions
             {
                 stream.Read(tempBuf, 0, 1);
 
-                blockLength = (int)tempBuf[0];
+                blockLength = tempBuf[0];
                 stream.Position += blockLength;
 
                 // Guard against infinite loop.
